[dev]: split code for smaller memory footprint
