+++ modules/lego_vehicle_timer.py
@@ -1,0 +1,405 @@
+# Timed train and vehicle program for interactive displays
+# Copyright Etendut
+# licence MIT
+
+from micropython import const
+from pybricks.parameters import Color, Side, Button
+from pybricks.pupdevices import Remote
+from pybricks.tools import wait, StopWatch
+#IMPORT_SECTION
+
+print('Version 1.4.0')
+##################################################################################
+#  Settings
+##################################################################################
+
+# countdown time settings
+COUNTDOWN_LIMIT_MINUTES: int = const(
+    3)  # run for (x) minutes, min 1 minute, max up to you. the default of 3 minutes is play tested :).
+# c = center button, + = + button, - = - button
+COUNTDOWN_RESET_CODE = 'c,c,c'  # left center button, center button, right center button
+
+#VARS_SECTION
+
+##################################################################################
+# ---------Main program below, editing should not be needed -------------
+
+# {Insert drive module here}
+
+class ErrorFlashCodes:
+    def __init__(self):
+        self.flash_count = 1  # Other errors
+
+    def set_error_no_motor_on_a(self):
+        self.flash_count = 2
+
+    def set_error_no_motor_on_b(self):
+        self.flash_count = 3
+
+    def set_error_no_remote(self):
+        self.flash_count = 4
+
+    def flash_error_code(self):
+        """
+            this flashes the remote led
+        """
+
+        global hub
+
+        for f in range(self.flash_count):
+            hub.light.on(Color.RED)
+            wait(350)
+            hub.light.on(Color.NONE)
+            wait(350)
+        if self.flash_count > 1:
+            wait(2000)
+
+
+class MotorHelper:
+
+    def __init__(self, supports_flip: bool, supports_homing: bool):
+        self.supports_flip = supports_flip
+        self.supports_homing = supports_homing
+
+    def handle_flip(self):
+        pass
+
+    def handle_homing(self):
+        pass
+
+    def handle_remote_press(self):
+        pass
+
+    def stop_motors(self):
+        pass
+
+
+
+
+##################################################################################
+# Countdown helper
+##################################################################################
+
+def wait_for_no_pressed_buttons():
+    buttons_pressed = remote.buttons.pressed()
+    while buttons_pressed:
+        buttons_pressed = remote.buttons.pressed()
+        wait(100)
+
+
+def convert_millis_hours_minutes_seconds(millis: int):
+    """
+        utility for making milliseconds hours/minutes/seconds
+    :param millis:
+    :return hours, minutes, seconds:
+    """
+    hours = int((millis / (1000 * 60 * 60)) % 24)
+    minutes = int((millis / (1000 * 60)) % 60)
+    seconds = int((millis / 1000) % 60)
+
+    return hours, minutes, seconds
+
+
+READY: int = const(0)
+ACTIVE: int = const(10)
+FINAL_MINUTE: int = const(20)
+FINAL_20_SECS: int = const(30)
+ENDED: int = const(40)
+
+
+class CountdownTimer:
+    """
+    This allows the model to run for a set time
+    """
+
+    def __init__(self):
+        # assign external objects to properties of the class
+        self.countdown_status = None
+
+        # Start a timer.
+        self.countdown_stopwatch = StopWatch()
+        self.led_flash_stopwatch = StopWatch()
+        self.last_console_msg = None
+        self.end_time = 0
+
+    def has_time_remaining(self):
+        """
+            Checks if countdown has time remaining
+        :return:
+        """
+        if self.countdown_status == ENDED or self.countdown_status == READY:
+            return False
+
+        # calculate remaining_time time
+        remaining_time = self.end_time - self.countdown_stopwatch.time()
+
+        # print a friendly console message
+        con_hour, con_min, con_sec = convert_millis_hours_minutes_seconds(int(remaining_time))
+        msg = 'countdown ending in: {}:{:02}'.format(con_min, con_sec)
+        if self.last_console_msg != msg:
+            self.last_console_msg = msg
+            print(msg)  # when time has run out end countdown
+        if remaining_time <= 0:
+            self.countdown_status = ENDED
+            self.show_status()
+            return False
+        # in last 25s slow flash a warning
+        if remaining_time < (1000 * 20):
+            self.countdown_status = FINAL_20_SECS
+            self.show_status()
+
+            # in last minute slow flash a warning
+        if remaining_time < (1000 * 60):
+            self.countdown_status = FINAL_MINUTE
+            self.show_status()
+
+        return True
+
+    def __start_countdown__(self):
+        """
+            start the countdown sequence by resetting timers and status
+        """
+        print('start countdown')
+        self.countdown_status = ACTIVE
+        self.show_status()
+        self.countdown_stopwatch.reset()
+        self.end_time = self.countdown_stopwatch.time() + (COUNTDOWN_LIMIT_MINUTES * 60 * 1000)
+
+    def reset(self):
+        print('countdown time reset, press Remote CENTER to restart countdown')
+        self.countdown_status = READY
+
+    def check_remote_buttons(self):
+        """
+            check countdown time buttons
+        """
+        pressed = remote.buttons.pressed()
+
+        if self.countdown_status == READY and Button.CENTER in pressed:
+            self.__start_countdown__()
+            wait_for_no_pressed_buttons()
+
+        # if reset sequence pressed reset the countdown timer
+        if all(i in pressed for i in PROGRAM_RESET_CODE_PRESSED) and not any(
+                i in pressed for i in PROGRAM_RESET_CODE_NOT_PRESSED):
+            self.reset()
+            wait_for_no_pressed_buttons()
+
+    def check_reset_code_pressed(self):
+        """
+           Check if reset code was pressed
+        """
+        pressed = remote.buttons.pressed()
+        # if reset sequence pressed at other times, end countdown
+        if self.countdown_status != READY and all(i in pressed for i in PROGRAM_RESET_CODE_PRESSED) and not any(
+                i in pressed for i in PROGRAM_RESET_CODE_NOT_PRESSED):
+            print('reset code pressed')
+            self.reset()
+            wait_for_no_pressed_buttons()
+
+    def show_status(self):
+        global hub
+        if self.countdown_status == READY:
+            self.__flash_remote_and_hub_light__(Color.GREEN, 500, Color.NONE, 500)
+        elif self.countdown_status == ACTIVE:
+            hub.light.on(Color.GREEN)
+            remote.light.on(Color.GREEN)
+        elif self.countdown_status == FINAL_20_SECS:
+            self.__flash_remote_and_hub_light__(Color.ORANGE, 200, Color.NONE, 100)
+        elif self.countdown_status == FINAL_MINUTE:
+            self.__flash_remote_and_hub_light__(Color.ORANGE, 500, Color.NONE, 250)
+        elif self.countdown_status == ENDED:
+            hub.light.on(Color.ORANGE)
+            remote.light.on(Color.ORANGE)
+
+    def __flash_remote_and_hub_light__(self, on_color, on_msec: int, off_color, off_msec: int):
+        """
+            this flashes the remote led
+        :param on_color:
+        :param on_msec:
+        :param off_color:
+        :param off_msec:
+        """
+        global hub
+        # we use a timer to make it a non-blocking call
+        if self.led_flash_stopwatch.time() > (on_msec + off_msec):
+            self.led_flash_stopwatch.reset()
+            remote.light.on(off_color)
+            hub.light.on(off_color)
+        elif self.led_flash_stopwatch.time() > off_msec:
+            remote.light.on(on_color)
+            hub.light.on(on_color)
+
+
+##################################################################################
+# Code helper
+##################################################################################
+
+def code_to_button_press_hash(button_code):
+    """
+    Returns the button needed to match a given code
+    :param button_code:
+    :return buttons_that_are_pressed[], buttons_that_should_not_pressed[]:
+    """
+    code_items = button_code.split(',')
+    buttons_that_should_not_pressed = [Button.LEFT_PLUS, Button.LEFT_MINUS, Button.LEFT, Button.CENTER,
+                                       Button.RIGHT_PLUS, Button.RIGHT_MINUS, Button.RIGHT]
+    buttons_that_are_pressed = []
+    # left code
+    if '+' in code_items[0]:
+        buttons_that_are_pressed.append(Button.LEFT_PLUS)
+        buttons_that_should_not_pressed.remove(Button.LEFT_PLUS)
+    if '-' in code_items[0]:
+        buttons_that_are_pressed.append(Button.LEFT_MINUS)
+        buttons_that_should_not_pressed.remove(Button.LEFT_MINUS)
+    if 'c' in code_items[0]:
+        buttons_that_are_pressed.append(Button.LEFT)
+        buttons_that_should_not_pressed.remove(Button.LEFT)
+
+    # middle code
+    if 'c' in code_items[1]:
+        buttons_that_are_pressed.append(Button.CENTER)
+        buttons_that_should_not_pressed.remove(Button.CENTER)
+
+    # right code
+    if '+' in code_items[2]:
+        buttons_that_are_pressed.append(Button.RIGHT_PLUS)
+        buttons_that_should_not_pressed.remove(Button.RIGHT_PLUS)
+    if '-' in code_items[2]:
+        buttons_that_are_pressed.append(Button.RIGHT_MINUS)
+        buttons_that_should_not_pressed.remove(Button.RIGHT_MINUS)
+    if 'c' in code_items[2]:
+        buttons_that_are_pressed.append(Button.RIGHT)
+        buttons_that_should_not_pressed.remove(Button.RIGHT)
+
+    return buttons_that_are_pressed, buttons_that_should_not_pressed
+
+
+PROGRAM_RESET_CODE_PRESSED, PROGRAM_RESET_CODE_NOT_PRESSED = code_to_button_press_hash(COUNTDOWN_RESET_CODE)
+
+
+##################################################################################
+# Main program
+##################################################################################
+class MockLight:
+    """For typing only, this will be replaced by CityHub or TechnicHub"""
+
+    def blink(self, col, seq):
+        pass
+
+    def on(self, color):
+        pass
+
+
+class MockIMU:
+    """For typing only, this will be replaced by TechnicHub"""
+
+    def up(self) -> Side:
+        pass
+
+
+class MockHub:
+    """For typing only, this will be replaced by CityHub or TechnicHub"""
+
+    def __init__(self):
+        self.imu = MockIMU()
+        self.light = MockLight()
+
+
+hub: MockHub = MockHub()
+
+
+def setup_hub():
+    global hub
+
+    try:
+        # this import will fail if the city hub is not connected.
+        from pybricks.hubs import CityHub
+        hub = CityHub()
+        print('Lego City Hub found')
+        return False
+    except ImportError as ex1:
+        print(ex1)
+        try:
+            from pybricks.hubs import TechnicHub
+            hub = TechnicHub()
+            print('Lego Technic Hub found')
+            return True
+
+        except ImportError as ex2:
+            print(ex2)
+            raise Exception('This program only support Lego City hub and Lego Technic hub')
+
+
+remote: Remote | None = None
+LED_FLASHING_SEQUENCE = [75] * 5 + [1000]
+
+
+def setup_remote(error_flash_code_helper, retry=5):
+    global hub
+    global remote
+
+    # Flashing led while waiting connection
+    hub.light.blink(Color.WHITE, LED_FLASHING_SEQUENCE)
+
+    # try to connect to remote multiple times
+    remote_retry_count = 1
+
+    while True:
+        # noinspection PyBroadException
+        try:
+            print("--looking for remote try " + str(remote_retry_count))
+            # Connect to the remote.
+            remote = Remote()
+            print("--remote connected.")
+            break
+        except:
+            if remote_retry_count >= retry:
+                error_flash_code_helper.set_error_no_remote()
+                raise  # ignore first 20 errors
+        remote_retry_count += 1
+        wait(50)
+
+#VEHICLE_SECTION
+
+def main():
+    error_flash_code = ErrorFlashCodes()
+    print('SETUP')
+    print('--setup hub')
+    try:
+        print("--setup countdown")
+        countdown_timer = CountdownTimer()
+        print("--setup motors")
+        drive_motors = MotorHelper(False, False)
+
+        print('--setup remote')
+        setup_remote(error_flash_code)
+
+        # give everything a chance to warm up
+        wait(500)
+
+        print('SETUP complete')
+
+        countdown_timer.reset()
+        while True:
+            countdown_timer.check_remote_buttons()
+            if countdown_timer.has_time_remaining():
+                if drive_motors.supports_flip:
+                    drive_motors.handle_flip()
+                drive_motors.handle_remote_press()
+            else:
+                drive_motors.stop_motors()
+                if drive_motors.supports_homing:
+                    drive_motors.handle_homing()
+
+            countdown_timer.show_status()
+            # add a small delay to keep the loop stable and allow for events to occur
+            wait(100)
+    except Exception as e:
+        print(e)
+        while True:
+            error_flash_code.flash_error_code()
+
+
+if __name__ == "__main__":
+    main()
